Flytta konverteringsfunksjoner og fjernet skip(1)
